pyproject.toml
[project]
name = "fastapi_api_financeira"
version = "0.1.0"
description = ""
authors = [
    {name = "Fillipe Berssot",email = "fillipeberssot95@hotmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "fastapi (>=0.121.2,<0.122.0)",
    "uvicorn[standard] (>=0.38.0,<0.39.0)",
    "sqlalchemy (>=2.0.44,<3.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "pydantic[email] (>=2.12.4,<3.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
ruff = "^0.14.5"
taskipy = "^1.14.1"
pytest = "^9.0.1"

[tool.ruff]
line-length = 88
target-version = "py312"
src = ["app"]
exclude = ["alembic"]

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.taskipy.tasks]
lint = "ruff check"
pre_format = "ruff check --fix"
format = "ruff format"
run = "uvicorn app.main:app --reload"
pre_test = "task lint"
test = "poetry run pytest"
init_db = "python -m app.core.init_db"


